Moved example pages to src
